Change rendering order to prevent visual issues.
According to the [SVG spec](svg-spec), subsequent elements are painted on top of
previously painted elements.

Because sparkline.js separated the fill and
the line (in order to only draw the top outline of the path), it makes
sense to draw the fill first and the line later. Otherwise, the fill is
shown over the line, causing stroke-width being cut in half and corners
looking off.

[svg-spec]: https://www.w3.org/TR/SVG/render.html#RenderingOrder